+++ tests/UKHO.S100PermitService.API.FunctionalTests/Configuration/TestConfiguration.cs
@@ -5,6 +5,10 @@ namespace UKHO.S100PermitService.API.FunctionalTests.Configuration
 {
     public static class TestConfiguration
     {
+        /// <summary>
+        /// This method is used to load the app settings configuration 
+        /// </summary>
+        /// <returns></returns>
         public static IConfigurationRoot LoadConfiguration()
         {
             var configBuilder = new ConfigurationBuilder()
@@ -14,6 +18,10 @@ namespace UKHO.S100PermitService.API.FunctionalTests.Configuration
             return configurationRoot;
         }
 
+        /// <summary>
+        /// This method is used to load and set the values of app settings configs. 
+        /// </summary>
+        /// <returns></returns>
         public static ServiceProvider ConfigureServices()
         {
             var services = new ServiceCollection();
