+++ tests/UKHO.S100PermitService.API.FunctionalTests/appsettings.json
@@ -4,7 +4,7 @@
     "InvalidToken": "TestInvalidToken",
     "ValidLicenceId": 2,
     "NonIntegerLicenceIds": [ "1A", "Test@123", "XYZ" ],
-    "InvalidHoldingsLicenceId": [ 0, 999 ],
+    "InvalidHoldingsLicenceId": [ 0, 999, -1 ],
     "InvalidUPNLicenceId": [ 0, 22 ],
     "InvalidPKSLicenceId": [1,3]
   },
